aw-models: Split library out from aw-server-rust
