+++ Cargo.toml
@@ -12,6 +12,7 @@ legacy_datastore_tests = []
 [workspace]
 members = [
     "aw-client-rust",
+    "aw-models",
 ]
 
 [lib]
@@ -47,6 +48,7 @@ toml = "0.5"
 gethostname = "0.2"
 regex = "1.0"
 uuid = { version = "0.8", features = ["serde", "v4"] }
+aw-models = { path = "aw-models" }
 
 [target.'cfg(target_os="android")'.dependencies]
 jni = { version = "0.5", default-features = false }
