+++ aw-models/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "aw-models"
+version = "0.1.0"
+authors = ["Johan Bjäreholt <johan@bjareho.lt>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+name = "aw_models"
+crate-type = ["lib", "cdylib"]
+path = "src/lib.rs"
+
+
+[dependencies]
+chrono = { version = "0.4", features = ["serde"] }
+log = "0.4"
+serde = "1.0"
+serde_json = "1.0"
+serde_derive = "1.0"
