+++ aw-models/src/timeinterval.rs
