+++ src/datastore/legacy_import.rs
@@ -7,9 +7,9 @@ use chrono::DateTime;
 use chrono::Utc;
 use chrono::Duration;
 
-use crate::models::Bucket;
-use crate::models::BucketMetadata;
-use crate::models::Event;
+use aw_models::Bucket;
+use aw_models::BucketMetadata;
+use aw_models::Event;
 
 use crate::datastore::DatastoreInstance;
