+++ src/endpoints/query.rs
@@ -3,9 +3,10 @@ use rocket::http::Status;
 use rocket::response::status;
 use rocket_contrib::json::{Json, JsonValue};
 
+use aw_models::Query;
+
 use crate::query;
 use crate::query::QueryError;
-use crate::models::Query;
 use crate::endpoints::ServerState;
 
 #[derive(Serialize)]
