--- src/models/mod.rs
@@ -1,13 +1,0 @@
-mod duration;
-mod bucket;
-mod event;
-mod timeinterval;
-mod query;
-
-pub use self::bucket::Bucket;
-pub use self::bucket::BucketMetadata;
-pub use self::bucket::BucketsExport;
-pub use self::event::Event;
-pub use self::timeinterval::TimeInterval;
-pub use self::query::Query;
-
