+++ tests/transform.rs
@@ -13,7 +13,8 @@ mod transform_tests {
     use chrono::Duration;
     use serde_json::json;
 
-    use aw_server::models::Event;
+    use aw_models::Event;
+
     use aw_server::transform;
 
     #[test]
